# pkg/db/dbif.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "pkg/db/dbif.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/dbif/GetTrials/{study_id}": {
      "get": {
        "summary": "* \nGet a Trial Configs from DB by ID of Study.",
        "operationId": "GetTrials",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/dbGetTrialsReply"
            }
          }
        },
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "DBIF"
        ]
      }
    },
...
  },
  "definitions": {
...
    "dbGetTrialsReply": {
      "type": "object",
      "properties": {
        "trials": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dbTrial"
          }
        }
      },
      "description": "*\nReturn a trial list in specified Study."
    },
...
    "dbParameter": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "parameter_type": {
          "$ref": "#/definitions/dbParameterType"
        },
        "value": {
          "type": "string"
        }
      },
      "description": "*\nValue of a Hyper parameter.\nThis will be created from a correcponding Config."
    },
...
    "dbParameterType": {
      "type": "string",
      "enum": [
        "UNKNOWN_TYPE",
        "DOUBLE",
        "INT",
        "DISCRETE",
        "CATEGORICAL"
      ],
      "default": "UNKNOWN_TYPE",
      "description": "*\nTypes of value for HyperParameter."
    },
...
    "dbTag": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      },
      "description": "*\nTag for each resource."
    },
    "dbTrial": {
      "type": "object",
      "properties": {
        "trial_id": {
          "type": "string"
        },
        "study_id": {
          "type": "string"
        },
        "parameter_set": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dbParameter"
          }
        },
        "objective_value": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dbTag"
          }
        }
      },
      "description": "*\nA set of Hyperparameter.\nIn a study, multiple trials are evaluated by workers.\nSuggestion service will generate next trials."
    },
...
  }
}